service/Services/entities/src/main/java/org/infy/idp/entities/jobs/basicinfo/Event.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.idp.entities.jobs.basicinfo;

import java.util.List;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

/**
 * Entity to store build event detail
 * 
 * @author Infosys
 *
 */
public class Event {
	@SerializedName("type")
	@Expose
	private String type;
	@SerializedName("hour")
	@Expose
	private String hour;
	@SerializedName("minute")
	@Expose
	private String minute;
	@SerializedName("date")
	@Expose
	private List<String> date;
	@SerializedName("week")
	@Expose
	private List<String> week;

	public String getHour() {
		return hour;
	}

	public void setHour(String hour) {
		this.hour = hour;
	}

	public String getMinute() {
		return minute;
	}

	public void setMinute(String minute) {
		this.minute = minute;
	}

	public List<String> getDate() {
		return date;
	}

	public void setDate(List<String> date) {
		this.date = date;
	}

	public List<String> getWeek() {
		return week;
	}

	public void setWeek(List<String> week) {
		this.week = week;
	}

	public String getType() {
		return type;
	}

	public void setType(String type) {
		this.type = type;
	}
}
